+++ app/src/main/java/com/stokerapps/chartmaker/ui/piechart/PieChartEntryItem.kt
@@ -5,11 +5,12 @@
 package com.stokerapps.chartmaker.ui.piechart
 
 import com.stokerapps.chartmaker.domain.PieChartEntry
+import java.math.BigDecimal
 
 data class PieChartEntryItem(
     val id: Int,
     var label: String,
-    var value: Float,
+    var value: BigDecimal,
     var color: Int
 ) {
     fun toDomainModel() = PieChartEntry(id, label, value, color)
